Initial commit
Added a README. I'm not exactly sure about the URL for the game yet,
but we'll see soon enough.

Did not add a license. The default when initializing with clj-new is
the Eclipse Public License, which seems fine, but I'm considering the
AGPL to prevent monetizing the game without supplying the source.

Adds configuration for deployment to application.garden, building and
testing locally with tools.build.